client-requirements.txt
pyautogui
requests
pillow